decision-making-engine/server/requirements.txt
Flask
flask_restful
Redis
scipy
numpy
matplotlib
argparse
igraph
graphviz
setuptools
uuid
psutil
freeport
requests
stomp-py==6.1.0
stompest
pycrypto
pycryptodome
Crypto
iso8601
pytz